policy: fold runtime reporting into governance core
