+++ Cargo.toml
@@ -7,7 +7,6 @@ members = [
   "crates/parsetable-metadata",
   "crates/bdd-grammar-fixtures",
   "crates/governance-runtime-core",
-  "crates/governance-runtime-reporting",
   "crates/governance-metadata",
   "crates/linecol-core",
   "crates/bdd-governance-core",
