--- crates/governance-runtime-reporting/Cargo.toml
@@ -1,33 +1,0 @@
-[package]
-name = "adze-governance-runtime-reporting"
-version = "0.1.0"
-edition.workspace = true
-rust-version.workspace = true
-license = "MIT OR Apache-2.0"
-description = "Shared reporting helpers for BDD progress and parser feature-profile diagnostics."
-repository = "https://github.com/EffortlessMetrics/adze"
-publish = ["crates.io"]
-keywords = ["governance", "parser", "metadata", "bdd", "policy"]
-categories = ["development-tools"]
-homepage = "https://github.com/EffortlessMetrics/adze"
-documentation = "https://docs.rs/adze-governance-runtime-reporting"
-
-[features]
-default = []
-pure-rust = ["adze-bdd-governance-core/pure-rust"]
-tree-sitter-c2rust = ["adze-bdd-governance-core/tree-sitter-c2rust"]
-tree-sitter-standard = ["adze-bdd-governance-core/tree-sitter-standard"]
-glr = ["pure-rust", "adze-bdd-governance-core/glr"]
-strict_api = ["adze-bdd-governance-core/strict_api"]
-strict_docs = ["adze-bdd-governance-core/strict_docs"]
-
-[dependencies]
-adze-bdd-governance-core = { version = "0.1.0", path = "../bdd-governance-core" }
-
-[dev-dependencies]
-adze-bdd-grid-core = { version = "0.1.0", path = "../bdd-grid-core" }
-adze-feature-policy-core = { version = "0.1.0", path = "../feature-policy-core" }
-adze-governance-metadata = { version = "0.1.0", path = "../governance-metadata" }
-adze-governance-runtime-core = { version = "0.1.0", path = "../governance-runtime-core" }
-proptest = { workspace = true }
-serde_json.workspace = true
