--- crates/governance-runtime-reporting/README.md
@@ -1,10 +1,0 @@
-# adze-governance-runtime-reporting
-
-Shared reporting helpers for BDD progress and parser feature-profile diagnostics.
-
-Part of the [adze](https://github.com/EffortlessMetrics/adze) workspace.
-
-## License
-
-Licensed under either of [Apache License, Version 2.0](../../LICENSE-APACHE)
-or [MIT License](../../LICENSE-MIT) at your option.
